+++ src/frontend/apps/e2e/__tests__/app-impress/doc-grid-dnd.spec.ts
@@ -1,6 +1,7 @@
 import { expect, test } from '@playwright/test';
 
 import { createDoc, mockedListDocs } from './utils-common';
+import { createRootSubPage } from './utils-sub-pages';
 
 test.describe('Doc grid dnd', () => {
   test('it creates a doc', async ({ page, browserName }) => {
@@ -162,6 +163,40 @@ test.describe('Doc grid dnd', () => {
     );
 
     await page.mouse.up();
+  });
+});
+
+test.describe('Doc grid dnd mobile', () => {
+  test.use({ viewport: { width: 500, height: 1200 } });
+
+  test('DND is deactivated on mobile', async ({ page, browserName }) => {
+    await page.goto('/');
+
+    const docsGrid = page.getByTestId('docs-grid');
+    await expect(page.getByTestId('docs-grid')).toBeVisible();
+    await expect(page.getByTestId('grid-loader')).toBeHidden();
+
+    await expect(docsGrid.getByRole('row').first()).toBeVisible();
+    await expect(docsGrid.locator('.--docs--grid-droppable')).toHaveCount(0);
+
+    await createDoc(page, 'Draggable doc mobile', browserName, 1, false, true);
+
+    await createRootSubPage(
+      page,
+      browserName,
+      'Draggable doc mobile child',
+      true,
+    );
+
+    await page
+      .getByRole('button', { name: 'Open the header menu' })
+      .getByText('menu')
+      .click();
+
+    await expect(page.locator('.--docs-sub-page-item').first()).toHaveAttribute(
+      'draggable',
+      'false',
+    );
   });
 });